Store per-edge values in the shared basis instead of padded planes
The backend rework left the tree half-ported: the new EnginePointer headers
were in place but k2tree.cpp, weight_matrix.cpp and engine.cpp still called
the deleted GpuPointer API, and nothing compiled.

Finishing it turned up the reason the old storage was so large. Every
per-edge value -- weight, delay, synapse prototype, synapse state -- lived
in a dense [node_count x max_neighbor_count] f32 plane, padded to the widest
row in the graph. set_edge_weight also pinned each matrix's Ck to zero, so
U*Ck*V was identically zero on the model path and the low-rank basis carried
no information about any edge at all. About 1.6 GB at a million neurons,
none of it compressed.

Per-edge storage is gone:

  - Synapse state collapses to one accumulator per (target, prototype).
    Every Phase-1 synapse is linear in its state with per-prototype
    parameters and weight entering the arrival increment linearly, so the
    sum of a target's incoming states obeys the same equation each term
    does. Exact, not approximate -- and an edge now touches memory only
    when a spike actually arrives rather than every tick.

  - Weight and delay come from the shared U/V basis, built directly from
    the projection structure rather than fitted. Give each projection run a
    latent lane, with U and V as 0/1 indicators and Ck holding the value;
    the reconstruction then collapses to exactly that run's value, because
    populations are disjoint so no other lane is nonzero at both endpoints.
    Measured against the model's own declarations afterwards regardless,
    since disjointness is a property of the model, not of the code.

  - Prototype index is a run table baked into the kernel. It picks a switch
    case, so it deliberately does not ride on a reconstruction.

  - Edges are numbered by a prefix sum over real out-degree, so nothing is
    sized by max_neighbor_count any more. It no longer sizes any allocation
    and is gone from the kernel entirely.

refit() and its hand-rolled Cholesky go with the dense Sk it fitted to:
with no stored per-edge values there is no ground truth to fit against.

Also here: partition/allocate is reusable, so SpikeEngine, WeightMatrix and
K2Tree each get their own slab from one backend; partitions no longer read
the slab pointer before allocate() creates it, and alignment is applied;
scalars reach kernels as EnginePointers bound with setBytes; the Metal
command queue and pipeline states are built once rather than per dispatch;
last_spiked uses a negative sentinel the refractory gate branches on instead
of a magic constant sized to outrun any refractory period.

Verified end to end. The two existing fixtures declare no synapses, so
two_cell_synapse.nml is new: two driven sources into one sub-rheobase
target, whose only route to threshold is the synapse. At zero weight the
target rests at -53.0007 mV against a hand-derived -53.0 mV; raising the
weights takes it to 0, 0, 0, 2, 3 spikes. Weights (1.0, 0.5) and (1.0, 1.0)
give different answers, which an implementation ignoring per-edge weight
could not. Rank derivation reports a worst weight error of exactly zero.

+++ include/spikecorec/nml/nml.h
@@ -479,7 +479,7 @@ struct NML_ParseResult {
 // Projection of neurons' outgoing edges into the plain source -> targets form WeightMatrix
 // consumes. Derived on demand rather than stored, so there is one edge list to keep
 // correct rather than two that can disagree.
-Vector<Vector<s64>> build_adjacency_list(const NML_ParseResult &parse_result);
+Vector<Vector<s32>> build_adjacency_list(const NML_ParseResult &parse_result);
 
 struct NML_Parser {
     UnorderedMap<String, ComponentType> declared_component_types;
